# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/production/step-logs": {
      "post": {
        "operationId": "StepLogsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStepLogDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "StepLogs"
        ]
      },
      "get": {
        "operationId": "ProductionController_getStepLogs",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Production"
        ]
      }
    },
...
    "/qc/audits": {
      "post": {
        "operationId": "QCAuditsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQCAuditDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "QCAudits"
        ]
      },
      "get": {
        "operationId": "QCAuditsController_findAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "QCAudits"
        ]
      }
    },
...
    "/production/machines": {
      "get": {
        "operationId": "ProductionController_getMachines",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Production"
        ]
      }
    },
...
    "/production/start/{workOrderId}": {
      "post": {
        "operationId": "ProductionController_startProduction",
        "parameters": [
          {
            "name": "workOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Production"
        ]
      }
    },
    "/production/{workOrderId}/submit-log": {
      "post": {
        "operationId": "ProductionController_submitLog",
        "parameters": [
          {
            "name": "workOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Production"
        ]
      }
    },
...
    "/production/requisitions/{id}/issue": {
      "post": {
        "operationId": "ProductionController_issueReq",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Production"
        ]
      }
    },
    "/production/requisitions/{id}/shortage": {
      "post": {
        "operationId": "ProductionController_shortageReq",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Production"
        ]
      }
    },
...
    "/production/qr/resolve/{uuid}": {
      "get": {
        "operationId": "ProductionController_resolveQR",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Production"
        ]
      }
    },
...
    "/production/schedules": {
      "post": {
        "operationId": "ProductionController_createSchedule",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Production"
        ]
      },
      "get": {
        "operationId": "ProductionController_getSchedules",
        "parameters": [
          {
            "name": "stage",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Production"
        ]
      }
    },
    "/production/schedules/{id}/result": {
      "post": {
        "operationId": "ProductionController_submitResult",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Production"
        ]
      }
    },
    "/production/schedules/{id}/actuals": {
      "post": {
        "operationId": "ProductionController_submitActuals",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Production"
        ]
      }
    },
...
    "/production/finalize/{woNumber}": {
      "post": {
        "operationId": "ProductionController_finalizeWorkOrder",
        "parameters": [
          {
            "name": "woNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Production"
        ]
      }
    },
    "/production/production-plans/{id}/assign-formula": {
      "post": {
        "operationId": "ProductionController_assignFormula",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Production"
        ]
      }
    },
...
    "/master/warehouses": {
      "get": {
        "operationId": "WarehousesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Warehouses"
        ]
      },
      "post": {
        "operationId": "WarehousesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWarehouseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Warehouses"
        ]
      }
    },
    "/master/warehouses/{id}": {
      "get": {
        "operationId": "WarehousesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Warehouses"
        ]
      },
      "patch": {
        "operationId": "WarehousesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWarehouseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Warehouses"
        ]
      },
      "delete": {
        "operationId": "WarehousesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Warehouses"
        ]
      }
    },
...
  },
  "info": {
    "title": "NexERP API",
    "description": "The ultimate manufacturing ERP system API documentation",
    "version": "4.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateStepLogDto": {
        "type": "object",
        "properties": {
          "workOrderId": {
            "type": "string",
            "format": "uuid"
          },
          "stage": {
            "type": "object"
          },
          "inputQty": {
            "type": "number"
          },
          "goodQty": {
            "type": "number"
          },
          "rejectQty": {
            "type": "number"
          },
          "quarantineQty": {
            "type": "number"
          },
          "machineId": {
            "type": "string",
            "format": "uuid"
          },
          "operatorId": {
            "type": "string",
            "format": "uuid"
          },
          "startTime": {
            "format": "date-time",
            "type": "string"
          },
          "endTime": {
            "format": "date-time",
            "type": "string"
          },
          "machineParams": {
            "type": "object"
          },
          "materialBatchNumber": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "workOrderId",
          "stage",
          "inputQty",
          "goodQty",
          "rejectQty",
          "quarantineQty"
        ]
      },
...
      "CreateQCAuditDto": {
        "type": "object",
        "properties": {
          "stepLogId": {
            "type": "string",
            "format": "uuid"
          },
          "inventoryId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "object"
          },
          "notes": {
            "type": "string"
          },
          "ph": {
            "type": "number"
          },
          "viscosity": {
            "type": "number"
          },
          "organoleptic": {
            "type": "string"
          },
          "fillingWeight": {
            "type": "number"
          },
          "sealingCheck": {
            "type": "string"
          },
          "labelingCheck": {
            "type": "string"
          },
          "expDateCheck": {
            "type": "string"
          },
          "halalStatus": {
            "type": "string"
          },
          "isCompliancePassed": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
...
      "CreateWarehouseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "picName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateWarehouseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "picName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      }
    }
  }
}